Refactor the accretion algorithm
The original implementation used the accrued mass from the latest iteration of the loop to alter the effect limit, which ignores the protoplanet mass.  This doesn't make a lot of difference with a new protoplanet, since generally the dust accrued is orders of magnitude heavier than the seed mass.  However, it leads to a worse accretion when re-evaluating a planet that has collided with a protoplanet,

This change changes this so that the mass of the protoplanet is included in the effect limit computations.